+++ test/olm/account_test.exs
@@ -1,0 +1,6 @@
+defmodule Olm.AccountTest do
+  use ExUnit.Case
+  alias Olm.Account
+
+  doctest Account
+end
